# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "this is an application web of interview assessment tests for interviewing out of the box .",
        "title": "test",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/my-tests/candidates/:test_id": {
            "post": {
                "description": "create test_candidate by json and path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test_candidate"
                ],
                "summary": "add new  test_candidate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Add candidate",
                        "name": "test_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Add candidate",
                        "name": "test_candidate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TestCandidate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TestCandidate"
                        }
                    }
                }
            }
        },
...
        "/quiz/currentQuestion/{idTestCandidate}": {
            "patch": {
                "description": "Update current question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update current question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "idTestCandidate",
                        "name": "idTestCandidate",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "current question",
                        "name": "currentQuestion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateCurrentQuestion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateCurrentQuestionOutput"
                        }
                    }
                }
            }
        },
        "/quiz/status/{idTestCandidate}": {
            "patch": {
                "description": "Update test status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update test status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "idTestCandidate",
                        "name": "idTestCandidate",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "test status",
                        "name": "testStatus",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTestStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTestStatusOutput"
                        }
                    }
                }
            }
        },
        "/quiz/{idTestCandidate}": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "get questions of a test",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test"
                ],
                "summary": "get tests candidates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "idTestCandidate",
                        "name": "idTestCandidate",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Question"
                            }
                        }
                    }
                }
            }
        },
        "/score/{idTestCandidate}": {
            "post": {
                "description": "calculate score by path  and update a status test",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test_candidate"
                ],
                "summary": "calculate a test score",
                "parameters": [
                    {
                        "type": "string",
                        "description": "idTestCandidate",
                        "name": "idTestCandidate",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TestCandidate"
                        }
                    }
                }
            }
        },
...
        "/startTest/{idTestCandidate}": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "get test information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test_candidate"
                ],
                "summary": "get test information",
                "operationId": "StartTest",
                "parameters": [
                    {
                        "type": "string",
                        "description": "idTestCandidate",
                        "name": "idTestCandidate",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.StartTest"
                            }
                        }
                    }
                }
            }
        },
...
        "/testsCandidates": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "get candidates and their tests",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test_candidate"
                ],
                "summary": "get candidates and their tests",
                "operationId": "TestsCandidatesList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "idTest",
                        "name": "idTest",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TestsCandidatesResponse"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Answer": {
            "type": "object",
            "required": [
                "question_id"
            ],
            "properties": {
                "answer_choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AnswerChoices"
                    }
                },
                "answer_file": {
                    "type": "string"
                },
                "answer_text": {
                    "type": "string"
                },
                "candidate_id": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "point": {
                    "type": "number"
                },
                "question_id": {
                    "type": "integer"
                },
                "test_id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.AnswerChoices": {
            "type": "object",
            "required": [
                "choices_id"
            ],
            "properties": {
                "answer_id": {
                    "type": "integer"
                },
                "choices_id": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Candidate": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "test": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Test"
                    }
                }
            }
        },
...
        "models.Choices": {
            "type": "object",
            "required": [
                "choice_text",
                "is_answer"
            ],
            "properties": {
                "choice_text": {
                    "type": "string"
                },
                "is_answer": {
                    "type": "boolean"
                }
            }
        },
...
        "models.Question": {
            "type": "object",
            "required": [
                "name",
                "question_text"
            ],
            "properties": {
                "answer": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Answer"
                    }
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Choices"
                    }
                },
                "difficulty": {
                    "type": "string"
                },
                "expected_time": {
                    "type": "integer"
                },
                "file_read_me": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "max_points": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "question_text": {
                    "type": "string"
                },
                "skill": {
                    "$ref": "#/definitions/models.Skill"
                },
                "skill_id": {
                    "type": "integer"
                },
                "test": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Test"
                    }
                },
                "test_questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TestQuestion"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Skill": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.StartTest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current_question": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StartTestQuestions"
                    }
                },
                "score": {
                    "type": "number"
                },
                "test_status": {
                    "type": "string"
                },
                "time_limit": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.StartTestQuestions": {
            "type": "object",
            "properties": {
                "expected_time": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Test": {
            "type": "object",
            "properties": {
                "archived": {
                    "type": "boolean"
                },
                "candidate": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Candidate"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "notify_emails": {
                    "type": "string"
                },
                "passing_score": {
                    "type": "integer"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Question"
                    }
                },
                "show_score": {
                    "type": "boolean"
                },
                "time_limit": {
                    "type": "integer"
                },
                "timing_policy": {
                    "type": "string"
                }
            }
        },
        "models.TestCandidate": {
            "type": "object",
            "required": [
                "candidate_id",
                "test_id"
            ],
            "properties": {
                "answer": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Answer"
                    }
                },
                "candidate_id": {
                    "type": "integer"
                },
                "current_question": {
                    "type": "integer"
                },
                "score": {
                    "type": "number"
                },
                "test_id": {
                    "type": "integer"
                },
                "test_status": {
                    "type": "string"
                }
            }
        },
        "models.TestQuestion": {
            "type": "object",
            "required": [
                "question_id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "question_id": {
                    "type": "integer"
                },
                "test_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.TestsCandidatesResponse": {
            "type": "object",
            "properties": {
                "candidate_email": {
                    "type": "string"
                },
                "candidate_name": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "test_candidate_id": {
                    "type": "string"
                },
                "test_name": {
                    "type": "string"
                },
                "test_status": {
                    "type": "string"
                }
            }
        },
        "models.UpdateCurrentQuestion": {
            "type": "object",
            "properties": {
                "current_question": {
                    "type": "integer"
                }
            }
        },
        "models.UpdateCurrentQuestionOutput": {
            "type": "object",
            "properties": {
                "current_question": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.UpdateTestStatus": {
            "type": "object",
            "properties": {
                "test_status": {
                    "type": "string"
                }
            }
        },
        "models.UpdateTestStatusOutput": {
            "type": "object",
            "properties": {
                "test_status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}